Add `ObservableObject::for_each()`
It's common to replicate the reaction code to be registered at an observable action like `after_change` as a `setup` to an initial value to `observed_t`. An observable action `for_each` can be used to act as an invariant to *each value* of the ObservableObject. Note: It is not supported to call `for_each` for a default constructed object.