Adds a namespace `view`
This changes isolates the artifacts that are related to the concept of
a view of an `ObservableObject` or an `ObservableErasableRange`
through a namespace `view`. It adds the directories `object/view` and
`container/view`. The `object_view<T>` and the `container_view<T>`
were renamed to `coruja::view::object<T>` and
`coruja::view::container<T>`. The class templates that allows type
erasure of an object or a container were renamed to
`coruja::view::any_object<T>` and `coruja::view::any_container<T>`.